add test cases for underlying function
